+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- CASMCMS-8271: Downloading recipe from S3 using boto3 package.
+
 ## [3.1.2] - 2024-09-19
 
 ### Changed